moving LoginButton into it's own folder
